fix(api call): fix all api call
