+++ app/front/src/components/list/listForum.js
@@ -1,16 +1,28 @@
-import React, {useState} from 'react';
+import React, {useEffect, useState} from 'react';
 import FiltersCard from "../filters/filtersCard";
-import CardForum from "../card/cardForum";
+import CardForum from "../card/forum/cardForum";
 import Button from "../button/button";
+import useForum from "../../hooks/api/useForum";
 
 
-const ListForum = ({text, list, setSidebar, setActiveSidebar}) => {
+const ListForum = ({text, setSidebar, setActiveSidebar}) => {
 
+    const [forums, setForums] = useState([]); // [state, function to update state
+    const {loading, error, handleGetAll} = useForum()
+    const [refresh, setRefresh] = useState(false); // [state, function to update state
     const [selectedForumId, setSelectedForumId] = useState(null);
 
     const handleForumSelect = (forumId) => {
         setSelectedForumId(forumId);
     };
+
+    useEffect(() => {
+        handleGetAll().then((response) => {
+            setForums(response);
+        }).catch((err) => {
+            console.log(err);
+        });
+    }, [refresh]);
 
     return (
         <div className="container-forum">
@@ -23,9 +35,9 @@ const ListForum = ({text, list, setSidebar, setActiveSidebar}) => {
                 <div className="container-forum__content">
 
                     {
-                        list.map((item, index) => {
+                        forums.map((item, index) => {
                             return(
-                                <CardForum key={index} forum={item} setSidebar={setSidebar} handleSelect={() => handleForumSelect(item.id)} isSelected={selectedForumId === item.id} setActiveSidebar={setActiveSidebar}/>
+                                <CardForum key={index} forum={item} setSidebar={setSidebar} handleSelect={() => handleForumSelect(item.id)} isSelected={selectedForumId === item.id} setActiveSidebar={setActiveSidebar} setRefresh={setRefresh} refresh={refresh}/>
                             )
                         }).reverse()
                     }
