--- app/front/src/components/list/listLibraryCompositors.js
@@ -1,30 +1,0 @@
-import React from 'react';
-import CardColumn from "../card/cardColumn";
-import FiltersCard from "../filters/filtersCard";
-
-const ListLibraryCompositors = ({list}) => {
-
-    return (
-        <div className="container-list">
-            <div className="list-row">
-                <div className="container__header">
-                    <h2>Récemment ajoutés</h2>
-                    <FiltersCard/>
-                </div>
-
-                <div className="container-list__content">
-
-                    {
-                        list.map((item, index) => {
-                            return(
-                                <CardColumn key={index} image={item.picture} title={item.name} subtitle={item.birth} description={""} link={`#/compositeurs/${item.id}`}/>
-                            )
-                        })
-                    }
-                </div>
-            </div>
-        </div>
-    );
-}
-
-export default ListLibraryCompositors;
