+++ app/front/src/components/list/listUsers.js
@@ -4,7 +4,7 @@ import Button from "../button/button";
 import edit from "../../assets/icones/icon-edit-Blue-stroke.svg";
 import trash from "../../assets/icones/icon-trash-White.svg";
 import PopupDelete from "../popup/popupDelete";
-import useUsers from "../../hooks/useUsers";
+import useUsers from "../../hooks/api/useUsers";
 
 const ListUsers= ({text}) => {
