+++ app/front/src/views/Authentification/Login.js
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from "react";
 import Input from "../../components/form/input";
 import logo from "../../assets/logo/logo-login.svg";
 import "../../styles/components/form.css";
-import useLogin from "../../hooks/useLogin";
+import useLogin from "../../hooks/api/useLogin";
 
 const Login = () => {
