+++ app/front/src/views/Student/Library/Oeuvres.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import MenuBar from "../../../components/navbar/MenuBar";
-import ListLibraryCompositors from "../../../components/list/listLibraryCompositors";
+import ListCompositors from "../../../components/list/listCompositors";
 import "../../../styles/library.css";
 import SidebarLibrary from "../../../components/sidebar/sidebarLibrary";
